refactor(utils): restructure browser utils
