+++ packages/lumx-react/src/hooks/useCallbackOnEscape.ts
@@ -1,6 +1,6 @@
 import { DOCUMENT } from '@lumx/react/constants';
 import { Callback } from '@lumx/react/utils/type';
-import { onEscapePressed } from '@lumx/react/utils/event';
+import { onEscapePressed } from '@lumx/react/utils/browser/event';
 import { useEffect } from 'react';
 import { Listener, makeListenerTowerContext } from '@lumx/react/utils/function/makeListenerTowerContext';
