+++ packages/lumx-react/src/utils/browser/isHoverNotSupported.ts
@@ -1,0 +1,2 @@
+/** Return true if the browser does not support pointer hover */
+export const isHoverNotSupported = (): boolean => !!window.matchMedia?.('(hover: none)').matches;
