+++ packages/lumx-react/src/utils/browser/isReducedMotion.ts
@@ -1,6 +1,6 @@
 import { WINDOW } from '@lumx/react/constants';
 
 /** Check if user prefers reduced motion */
-export function getPrefersReducedMotion() {
+export function isReducedMotion() {
     return WINDOW?.matchMedia?.('(prefers-reduced-motion: reduce)').matches;
 }
