--- packages/lumx-react/src/utils/browserDoesNotSupportHover.ts
@@ -1,2 +1,0 @@
-/** Return true if the browser does not support pointer hover */
-export const browserDoesNotSupportHover = (): boolean => !!window.matchMedia?.('(hover: none)').matches;
