--- packages/lumx-react/src/utils/userHasReducedMotion.ts
@@ -1,7 +1,0 @@
-export const userHasReducedMotion = () => {
-    try {
-        return window.matchMedia('(prefers-reduced-motion: reduce)').matches;
-    } catch (e) {
-        return false;
-    }
-};
